pyproject.toml
[project]
name = "project-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain-text-splitters>=1.1.1",
    "litellm>=1.81.13",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "openrouter>=0.6.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
]
